% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_theme.R
\name{my_theme}
\alias{my_theme}
\title{Custom ggplot2 theme}
\usage{
my_theme()
}
\value{
A customized ggplot2 theme with matching fill and color scales.
}
\description{
A custom theme for ggplot2 plots.
}
\examples{
library(ggplot2)

ggplot(mtcars, aes(x = factor(cyl), y = mpg, fill = factor(cyl))) +
  geom_col() +
  my_theme()

}
