% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean a Data Set}
\usage{
clean(data_frame, remove_missing = TRUE, remove_duplicate = TRUE)
}
\arguments{
\item{data_frame}{A data frame that needs to be cleaned for easier use}

\item{remove_missing}{A logical argument that tells the function to remove incomplete cases}

\item{remove_duplicates}{A logical argument that tells the function to remove duplicate cases}
}
\value{
The original data frame after being cleaned
}
\description{
Clean a Data Set
}
\examples{
data <- data.frame(
  Name = c("Taylor", "Taylor", "Adele"),
  Score = c(10, 10, NA)
)
clean(data, remove_missing = FALSE, remove_duplicate = FALSE)
clean(data, remove_missing = FALSE)
clean(data, remove_duplicate = FALSE)
clean(data)

}
